fixed extra table in session file
